Changed DefaultBalance.py to just contain a static method so objects don't have to be created for it. Also moved DefaultBalance.py to Fundamentals
