+++ src/pyacty/financial_statements/__init__.py
@@ -1,4 +1,3 @@
 from .BalanceSheet import *
-from .DefaultBalance import *
 from .FinancialStatement import *
 from .IncomeStatement import *
